requirements.txt
# Surface Defect Detector Requirements
ultralytics==8.3.40          # YOLOv8
torch==2.8.0+cpu             # PyTorch CPU
torchvision==0.19.0+cpu      # Vision utils
opencv-python==4.10.0.84     # CV2
streamlit==1.38.0            # Web app
pillow==10.4.0               # Image processing
numpy==1.26.4                # Arrays
tqdm==4.66.5                 # Progress bars